move drawer up
